+++ src/Component/Common/Modal.js
@@ -1,5 +1,6 @@
 import React from "react";
-import styles from '../assets/style/Modal.module.scss';
+import styles from '../../assets/style/Modal.module.scss';
+
 const Modal=( {isOpen , onClose , children}) => {
     if(!isOpen) return null;
   return(
